buffile: combine FileWrite for compressed buffile.
Add a small buffer to each compressed buffile, to combine FileWrite (the pwrite syscall).
the compressed data will copy to this small buffer first, then written to the physical file.

This can improve the buffile performance on some very fast (>500MiB/s) disks.

current the IO size is the default of ZSTD Level 1's window size, which is 1KiB.
test the performance with fio:

    fio -filename t -name=t -iodepth 1 -size 1G -bs 1k  -rw=readwrite -> READ 329MB/s  WRITE 329MB/s
    fio -filename t -name=t -iodepth 1 -size 1G -bs 32k -rw=readwrite -> READ 1307MB/s WRITE 1306MB/s

the patch also ensures the start offset of IO operation is aligned with BLCKSZ.
maybe we can do other features on top of this.

Signed-off-by: Sasasu <i@sasa.su>